clockify widgets are in one file
